+++ src/utils/botUtilsConstants.js
@@ -27,6 +27,8 @@ const HELP_MESSAGE = 'This is what I can do: \n'
     + '* `pay <username> <amount>`: I\'ll pay another player from your wallet\n'
     + '* `xp query <username>`: I\'ll tell you how much XP a player has\n'
     + '* `xp deposit <amount>`: I\'ll deposit your XP points into your XP bank\n'
+    + '* `xp withdraw <amount>`: I\'ll withdraw your XP points from your XP bank\n'
+    + '* `xp balance <username>`: I\'ll check your XP bank balance\n'
     + '* `help`: Show this help, since as it stands you can\'t do anything for yourself';
 
 const EMPTY_SEARCH_RESULTS_MESSAGE = "No results!";